build: add critools project for clone
And also check existing folder. If it is git folder there, skip it to
make sure that clone_github_projects can be run after vagrant up without
impacting existing works.

+++ build/clone_github_projects.sh
@@ -20,6 +20,7 @@ fetch_go_projects() {
     "github.com/containerd/containerd:fuweid/containerd"
     "github.com/moby/buildkit:fuweid/buildkit"
     "github.com/containerd/cri:fuweid/cri"
+    "github.com/kubernetes-sigs/cri-tools:fuweid/cri-tools"
   )
 
   # FIXME(fuweid): support no upstream type
@@ -40,6 +41,9 @@ fetch_go_projects() {
     fi
 
     # NOTE(fuweid): remove the existing dir, which only show up testing
+    if [[ -d ${pkg_dir}/.git ]]; then
+      continue
+    fi
     rm -rf "${pkg_dir}"
 
     # for me
